修复 .NET 10 ToString 链路所需运行时基建
补齐 .NET 10 CoreLib 在数值和装箱 ToString 路径上会触发的基础运行时入口：新增 Kernel32 环境变量与区域信息桥接、Marshal Get/SetLastPInvokeError 兼容入口、Advapi32/EventPipe 诊断事件空实现、Environment.GetProcessorCount 与 NtDll.RtlGetVersion stub，避免 CoreLib 初始化链路在诊断和环境查询上中断。

补充 Unsafe ByteOffset/CopyBlock/ReadUnaligned/WriteUnaligned/AsRef 等 intrinsic，并修正 RuntimeHelpers 布尔返回写入为 int32，修复嵌套 Runtime.Intrinsics IsSupported 判断，使 .NET 10 数值格式化、泛型数学与装箱 ToString 能稳定运行。

已验证：RunBox、RunBoxingInstructions、RunMemoryLdc 均可通过 scripts\\dotnet10\\interp-smoke.ps1 的 Release 冒烟；并清理了临时 trace 输出，未提交 IDE、artifacts、bin、obj 等无关噪声。

+++ src/runtime/intrinsics/system_runtime_compilerservices_runtimehelpers.cpp
@@ -225,7 +225,7 @@ static RtResultVoid is_bitwise_equatable_invoker(metadata::RtManagedMethodPointe
 
     DECLARING_AND_UNWRAP_OR_RET_ERR_ON_FAIL(bool, result,
                                             SystemRuntimeCompilerServicesRuntimeHelpers::is_bitwise_equatable(method));
-    interp::EvalStackOp::set_return(ret, result);
+    interp::EvalStackOp::set_return(ret, static_cast<int32_t>(result));
     RET_VOID_OK();
 }
 
@@ -238,7 +238,7 @@ static RtResultVoid is_reference_or_contains_references_invoker(metadata::RtMana
 
     DECLARING_AND_UNWRAP_OR_RET_ERR_ON_FAIL(bool, result,
                                             SystemRuntimeCompilerServicesRuntimeHelpers::is_reference_or_contains_references(method));
-    interp::EvalStackOp::set_return(ret, result);
+    interp::EvalStackOp::set_return(ret, static_cast<int32_t>(result));
     RET_VOID_OK();
 }
